Refactored file structure to accommodate for multiple pages, fixed kim's widget display issue, passing state up to main to display home or splash page based on if wifi is connected or not!
