+++ lib/screens/main/detections_page.dart
