--- lib/screens/wifi_page.dart
@@ -1,22 +1,0 @@
-import 'package:flutter/material.dart';
-
-class WifiPage extends StatelessWidget {
-  const WifiPage({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: Container(
-        decoration: const BoxDecoration(
-          image: DecorationImage(
-              image: AssetImage("assets/images/background2.JPG"),
-              fit: BoxFit.cover),
-        ),
-      
-      child: const Center(
-        child: Text("Wifi Page Content"),
-        ),
-      )
-    );
-  }
-}
